requirements.txt
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.16
django-crispy-forms==1.7.2
django-grappelli==2.10.2
django-rest-action-permissions==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
